requirements.txt
albumentations==1.3.1
asttokens==2.4.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
Cython==3.0.2
debugpy==1.8.0
decorator==5.1.1
easydict==1.10
executing==1.2.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.42.1
humanfriendly==10.0
idna==3.4
imageio==2.31.3
insightface==0.7.3
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.23.0
onnx==1.14.1
onnxruntime-gpu==1.16.0
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prettytable==3.9.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
qudida==0.0.4
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.9.18
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
wcwidth==0.2.6
Werkzeug==3.0.0
